Fixed CSS code processing and minification
The mrclay/minify extension did not handle relative paths correctly, in particular those with ../ and ./. It has written its own function that adjusts paths based on the initial location of the file and the final one.

The natxet/cssmin extension does not correctly process already minified code, so it was replaced with the matthiasmullie/minify extension.

Fix #43